Check availability of local storage
